Setup eth blockminer prototype
This commit creates a first draft of a prototype which should extract in a continuous manner the ethereum blockchain, both transactions and block metadata.

This initial service is creating a datadump of files both locally and on an s3 bucket. The files are being exported both in JSON and CSV format. This may change in the future.

Also the intial sql scripts that create the datawarehouse, schema, table and loading logic have been added aswell into the sql/ folder.

Currently the service is not running constantly. It only extracts one block at a time. A service must be created so that it continually checks for new blocks, and executes if a new block is found. For this it would be important to have a light database, such as sqllite, which indexes the blocks that have been mined.